+++ board/recalbox/fsoverlay/recalbox/scripts/recalbox-config.sh
@@ -485,6 +485,10 @@ EOF
 	done
 
         connmanctl enable wifi || exit 1
+	settings_region="`$systemsetting  -command load -key wifi.region`"
+	if [[ "$settings_region" != "" ]] ;then
+		/usr/sbin/iw reg set "${settings_region}"
+	fi
         connmanctl scan   wifi || exit 1
         exit 0
     fi
@@ -496,6 +500,10 @@ EOF
             fi
         fi
         connmanctl enable wifi || exit 1
+	settings_region="`$systemsetting  -command load -key wifi.region`"
+	if [[ "$settings_region" != "" ]] ;then
+		/usr/sbin/iw reg set "${settings_region}"
+	fi
         connmanctl scan   wifi || exit 1
         exit 0
     fi
@@ -504,6 +512,10 @@ EOF
         exit $?
     fi
     if [[ "$mode" == "list" ]]; then
+	settings_region="`$systemsetting  -command load -key wifi.region`"
+	if [[ "$settings_region" != "" ]] ;then
+		/usr/sbin/iw reg set "${settings_region}"
+	fi
         connmanctl scan wifi &> /dev/null
         WAVAILABLE=$(connmanctl services | grep "wifi_" | grep -v "_hidden_" | cut -c 5- | sed -e "s/\(.*[^ ]\) *wifi_.*/\1/g")
         if test -n "${ssid}"